Streamlied FetchTorrent() function so it doesn't reuse the same cut/pasted code over and over again to check if content is a torrent.
Add more meaningful messages when problems occur with fetching torrent data - print the first 1024 chars of the content returned if debuglevel is set > 1.  This may need a tweak, for example if the content is HTML this could break the webapp view if not escaped.  Perhaps using a textarea might make displaying this content better?