Initial commit. Wrote a rough draft/list/skeleton of contents (308 lines!).  Aim is to be exhaustive in terms of operating and configuring tfb.  Will flesh out the skeleton over time.
The file will be monolithic to start with, in plain text, but when
complete it could be marked up as XML/SGML(docbook?).  The XML could
then be parsed and content generated appropriately depending on the
target:

ie:
- for a monolithic txt file strip all markup completely and write one
  single .txt file called 'manual.txt'

- for frontend 'integrated' help documentation, create multiple small
  html files each containing the relevant documentation for a specific
  part of the web app (or create a PHP based help framework?)

- for generic html documentation suitable to view as a 'handbook',
  create one content page linking to several individual html files, one
  for each 'chapter' or section (user options, admin options, advanced
  options, troubleshooting) or create one single monolithic HTML
  document containing everything.

Just some ideas - not sure if writing the document or marking it up as
XML and all the other cruft is harder! :)