b4rt (or anyone else who has opera to hand) - if you read this can you check that the sizing of the two cells in the top part of the frontend look ok in opera?  I checked in ff/ie and it looks ok, tested with a few different options. --
Adjusted width of links/users/goodlookingstats table to fit better.  New
width is 425px - approx width of urlupload/wget/etc table is ~ 450px, so
425+450+somepadding=~900px is about right for default width.

Only issue is if usernames are long (over 10chars or so) or links
included are long (which is the case at the moment anyway so it's not a
problem introduced with this minor change).

A way around this would be to either:

- use the overflow css property to include scrollbars on the td's that
  contain each of those items, but scrollbars probably take up too much
  room there anyway.

- truncate the strings in those cells to some finite length, but this
  also is problematic for readability.