Added helper text to superadmin pages. No changes to functionality, only to readability / usability / layout / user feedback.
NOTE/TODO: I noticed the 'lock' feature was broken when I tested it.
Didn't get a chance to look into it, seems to be that the
$cfg['webapp_locked'] value isn't reported correctly around line 635 -
it can be set to locked in the db, but in the script it's reported as
not locked, cfg vbl doesn't seem to be set properly for some reason.