Added a short 'Overview:' section to the top of the stats admin page since it's not obvious from the content of the page what exactly the page refers to!!!
This might be an idea also for some of the other more obscure pages,
such as fluxd/links/rss/etc just to explain what exactly it is that the
items on each page are in reference to.  Maybe even worth having this
kind of overview section for every single admin page just to clarify
what's what.  Will see.