stats.php: bugfix: Fix to display usage when stats.php called with no _GET args and autologin cookie is being used.
If stats.php is called without args and $cfg[stats_show_usage] is set,
stats.php is supposed to display stats.php usage info.

However the script checks to see if count($params) == 0 for this, but
count($params) is always > 0 if an autologin cookie is used.

Instead, used count($_GET) instead of count($params).