Extended Trigger's flexibility. It can now act when a transfer is: started stopped completed seeding
Also, Trigger now has the skeleton to perform the following actions:
email user
run external script
move files
unzip file

NB: currently this is tested and appears to work for me, but it's
not ready to be used in production. Also, run external script is
the only action that currently works.